requirements.txt
azure-storage-blob==12.16.0
blinker==1.6.2
click==8.1.3
Flask==2.3.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
pandas==2.0.2
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
tzdata==2023.3
Werkzeug==2.3.4
weasyprint==59.0